requirements.txt
# Telegram Bot Framework
python-telegram-bot[ext]

# Database (ORM and Migrations)
sqlalchemy
alembic
psycopg2-binary  # For PostgreSQL, use 'mysqlclient' for MySQL or remove for SQLite

# HTTP Requests for APIs (Marzban, Payment Gateway)
requests
httpx

# Environment Variable Management
python-dotenv

# For handling dates and times
pytz

# For generating QR Codes
qrcode[pil]